Used a fixed-length array to represent the interpreter stack
CPython represents the stack of a frame as a fixed length C array + a stack pointer. The pointer points to the next free slot. When the interpreter executes a PUSH or a POP operation, it increments or decrements the pointer, without having to resize the underlying array. This allows fast O(1) append/pop operations.

ghstack-source-id: fe16a1351b5439eaba0ba0bc1da439beb62ef9c5
Pull-Request: https://github.com/pytorch/pytorch/pull/162724